Freeze Bug
I tried adding in walls that the snake could collide with, which worked, however every now and then food didn't seem to spawn. I think it was spawning beneath the wall objects, So i tried to add a condition to the spawn food function to make it so that food could not spawn on Snake Head, Snake Body or wall sprites.